Split smoke.rs into themed integration-test files
The single tests/smoke.rs (4818 lines, 61 tests) is split by command
surface so each area lives in its own file. Shared helpers move to
tests/common/mod.rs (made pub, #![allow(dead_code)] since each test
binary uses a different subset).

New files: project, bundle, merge, config, feature_flow, sync,
publish, land, cleanup. No test logic changed; only relocated.
